backend/requirements.txt
Flask==3.0.3
Flask-PyMongo==2.3.0
Flask-Bcrypt==1.0.1
Flask-JWT-Extended==4.6.0
python-dotenv==1.0.1
pymongo==4.8.0
ipfshttpclient==0.7.0
qrcode==7.4.2
Pillow==10.4.0
py-algorand-sdk==2.6.1
flask-cors==4.0.1
